flask test, create config.py
