+++ categorize.py
@@ -1,23 +1,15 @@
 import re
+from flask import Flask, render_template, redirect, request
 import tweepy
 import fasttext as ft
 import MeCab
+import config
 
 COUNT = 700    # ツイート取得数
 model = ft.load_model('/Users/fukunagaatsushi/Documents/gitdev/CategorizeTweets/model.bin')  # 分類器
 
-# 認証に必要なキーとトークン
-API_KEY = 'IR2CAa7c3w5OqzilT5iCPIAbg'
-API_SECRET = 'swUFZpyRUnzfYWOnRkfAbSRa2xZycAUTot5ype3j6czTU17dHY'
-ACCESS_TOKEN = '911457413865152512-kzcZi6uMkK8LHHxbCSXz5D23hPiNKv3'
-ACCESS_TOKEN_SECRET = '7CuJBoA7FtHdngpAzUQpaTLTEkPRhf4Td73WWRwzXtCwb'
-
-# TwitterAPI認証用関数
-def authTwitter():
-    auth = tweepy.OAuthHandler(API_KEY, API_SECRET)
-    auth.set_access_token(ACCESS_TOKEN, ACCESS_TOKEN_SECRET)
-    api = tweepy.API(auth)  #APIインスタンスの作成
-    return api
+# flask初期設定
+app =  Flask(__name__)
 
 def main():
     id = get_id()    # Twitter ID取得
@@ -32,7 +24,7 @@ def get_id():
 
 # 指定したユーザーのツイートを取得
 def get_tweet(id):
-    api = authTwitter() # 認証
+    api = config.authTwitter() # API認証
     tweets = [
         tweet
         for tweet in tweepy.Cursor(api.user_timeline, screen_name=id).items(COUNT)
@@ -83,5 +75,10 @@ def categorize(results, raw_tweets):
             print(category_dic[key][0])
             print('--------------------')
 
-if __name__ == '__main__':
-    main()
+
+@app.route('/')
+def index():
+    return render_template('index.html')
+
+if __name__ == "__main__":
+    app.run(debug=True)
