+++ config.py
@@ -1,0 +1,14 @@
+import tweepy
+
+# 認証に必要なキーとトークン
+API_KEY = 'IR2CAa7c3w5OqzilT5iCPIAbg'
+API_SECRET = 'swUFZpyRUnzfYWOnRkfAbSRa2xZycAUTot5ype3j6czTU17dHY'
+ACCESS_TOKEN = '911457413865152512-kzcZi6uMkK8LHHxbCSXz5D23hPiNKv3'
+ACCESS_TOKEN_SECRET = '7CuJBoA7FtHdngpAzUQpaTLTEkPRhf4Td73WWRwzXtCwb'
+
+# TwitterAPI認証用関数
+def authTwitter():
+    auth = tweepy.OAuthHandler(API_KEY, API_SECRET)
+    auth.set_access_token(ACCESS_TOKEN, ACCESS_TOKEN_SECRET)
+    api = tweepy.API(auth)  #APIインスタンスの作成
+    return api
