+++ templates/index.html
@@ -1,0 +1,5 @@
+{% extends 'base.html' %}
+
+{% block body %}
+<h1>extended!!!</h1>
+{% endblock %}
