+++ tweet_get.py
@@ -1,25 +1,13 @@
 import re
-import tweepy
 import MeCab
 from time import sleep
+import config
 
 COUNT = 100    # ツイート取得数
 train_idx = 80    # トレーニングデータ%
 SET = 30    # セット数×100件のツイートを取得
 KEYWORDS = ['エンジニア', 'ゲーム', 'アニメ']
 
-# 認証に必要なキーとトークン
-API_KEY = 'IR2CAa7c3w5OqzilT5iCPIAbg'
-API_SECRET = 'swUFZpyRUnzfYWOnRkfAbSRa2xZycAUTot5ype3j6czTU17dHY'
-ACCESS_TOKEN = '911457413865152512-kzcZi6uMkK8LHHxbCSXz5D23hPiNKv3'
-ACCESS_TOKEN_SECRET = '7CuJBoA7FtHdngpAzUQpaTLTEkPRhf4Td73WWRwzXtCwb'
-
-# TwitterAPI認証用関数
-def authTwitter():
-    auth = tweepy.OAuthHandler(API_KEY, API_SECRET)
-    auth.set_access_token(ACCESS_TOKEN, ACCESS_TOKEN_SECRET)
-    api = tweepy.API(auth)  #APIインスタンスの作成
-    return api
 
 def main():
     tweets = get_tweet()      #ツイートを取得
@@ -28,7 +16,7 @@ def main():
 
 # TwitterからKEYWORDに関連するツイートを取得
 def get_tweet():
-    api = authTwitter() # 認証
+    api = config.authTwitter() # API認証
     tweets = api.search_tweets(q=KEYWORD, count=COUNT)
     return tweets
